+++ docroot/modules/humsci/hs_actions/src/Plugin/Action/FieldClone/FieldCloneBase.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace Drupal\hs_actions\Plugin\Action\FieldClone;
+
+use Drupal\Core\Form\FormStateInterface;
+use Drupal\Core\Plugin\PluginBase;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+
+/**
+ * Base field for all field clone plugins.
+ *
+ * @package Drupal\hs_actions\Plugin\Action\FieldClone
+ */
+abstract class FieldCloneBase extends PluginBase implements FieldCloneInterface {
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
+    return new static($configuration, $plugin_id, $plugin_definition);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function buildConfigurationForm(array $form, FormStateInterface $form_state) {
+
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function validateConfigurationForm(array &$form, FormStateInterface $form_state) {
+
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function submitConfigurationForm(array &$form, FormStateInterface $form_state) {
+
+  }
+
+}